Factor out Select components
